Create Chat Vue component
